+++ src/dipdup/datasources/tzkt/datasource.py
@@ -12,6 +12,7 @@ from aiosignalrcore.transport.abstract import ConnectionState  # type: ignore
 from dipdup.config import (
     BigMapIndexConfig,
     ContractConfig,
+    HeadIndexConfig,
     HTTPConfig,
     OperationHandlerOriginationPatternConfig,
     OperationIndexConfig,
@@ -499,6 +500,10 @@ class TzktDatasource(IndexDatasource):
                     self._big_map_subscriptions[address] = set()
                 if path not in self._big_map_subscriptions[address]:
                     self._big_map_subscriptions[address].add(path)
+
+        # NOTE: head subscription is enabled by default
+        elif isinstance(index_config, HeadIndexConfig):
+            pass
 
         else:
             raise NotImplementedError(f'Index kind `{index_config.kind}` is not supported')